Remove the broken (and unneeded) no-cache header detection
The hook was registered to the 'tslib_fe-PostPrc' callback.
That callback is run at the end of the TSFE's constructor.
At that time the beUserLogin is not yet filled, but always initialized to NULL.
Cause of that the cache detection will always return FALSE and the hook never clears anything.

Removing this now since we're receiving a cache update in the insertPageIncache anyway
and don't need to check for the headers ourselves.