htaccess: Avoid 403 error because of the internal redirect to typo3temp...
Cause of the [L] flag on our last rewrite rule apache performs an internal redirect
to the rewritten path and goes through all rewrite condition's and rules again.
The typo3temp/tx_ncstaticfilecache direct access protection is triggered in that case.

To fix that, add a condition to the access check to match only for non-internal redirects.

Also add comments on how to perform better on Apache 2.4 by using [END] instead of [L]